Fix order of insight histories
The insight pump only supports second precision so when a TBR is changed, this is a quick succession of "cancel TBR" + "set TBR" in the same second. The code sorts the events multiple times and processes the events from new to old for some reason. This leads to the "set TBR" record to be inserted first sometimes and then the database will update the "set TBR" record to be the "cancel TBR" record since they happened in the same second so we can't rely on timestamp to tell us if one event happened before another. Luckily the insight has a monotonically increasing eventPosition id we can sort by to have the events in order of happening. By just avoiding sorting all over the place we can keep and process the events from old to new. We still override the "cancel TBR" with the "set TBR" with this fix but this does not appear to be a problem.

Additionally, i've noticed that the historyEvents curiously enough had milliseconds attached to them which were a random artifact of how the timestamp were computed.